ResNet50

+++ com/leo/koreanparser/main.py
@@ -2,11 +2,11 @@ import string
 import time
 import argparse
 from watchdog.observers import Observer
-from watchdog.events import FileSystemEventHandler, FileSystemEvent, EVENT_TYPE_CLOSED
+from watchdog.events import FileSystemEventHandler, FileSystemEvent, EVENT_TYPE_CLOSED, EVENT_TYPE_MODIFIED
 import os
 from dotenv import load_dotenv
 import cv2
-import sys, traceback, shutil, glob
+import sys, traceback, shutil
 import pandas as pd
 import numpy as np
 import torch
@@ -54,24 +54,27 @@ class Handler(FileSystemEventHandler):
         self.ensure_dir(work_directory)
         self.threshold = 0.5
         self.model = get_model(eval=True)
+        self.treated = set([])
         do_lod_specific_model(weights_path, self.model)
-        print(f"Work dir is {work_directory}")
 
     def ensure_dir(self, file_path):
-        print(f"Creating dir {file_path}")
         os.makedirs(file_path, exist_ok=True)
 
     def on_any_event(self, event: FileSystemEvent):
         ready_file_path = event.src_path
-        if not event.is_directory and event.event_type == EVENT_TYPE_CLOSED and ready_file_path and ready_file_path.endswith('.ready'):
+        if not event.is_directory and \
+                (event.event_type == EVENT_TYPE_CLOSED or event.event_type == EVENT_TYPE_MODIFIED) \
+                and ready_file_path and ready_file_path.endswith('.ready') and not ready_file_path in self.treated:
+            self.treated.add(ready_file_path)
             file_path = ready_file_path[:-6]
-            os.remove(ready_file_path)
             try:
                 self.treat_incoming_file(file_path)
             except:
                 print("-"*60)
                 traceback.print_exc(file=sys.stdout)
                 print("-"*60)
+            finally:
+                os.remove(ready_file_path)
 
     def treat_incoming_file(self, file_path):
         print(f"Clean working directory {self.work_directory}")
@@ -83,7 +86,8 @@ class Handler(FileSystemEventHandler):
                     os.remove(entry.path)
         print(f"Treating file {file_path}")
         prefix_splitted: str = self.split_file(file_path)
-        annotation_file = self.create_annotations(prefix_splitted)
+        annotation_file, annotations = self.create_annotations(prefix_splitted)
+        #self.extract_bounding_rects =
 
     def split_file(self, file_path) -> str:
         print(f"Splitting file {file_path}")
@@ -107,37 +111,45 @@ class Handler(FileSystemEventHandler):
         return filename_without_extension
 
     def create_annotations(self, prefix_splitted: str):
-        splitted_files = glob.glob(f"{self.work_directory}/{prefix_splitted}-*.jpg")
         size = (TARGET_WIDTH, TARGET_HEIGHT)
         data = []
         i = 0
-        for splitted_file in splitted_files:
-            print(f"--- Reading splitted file {splitted_file}")
-            im = read_image(splitted_file)
-            print(f"------ Resizing")
-            im = cv2.resize(im, size)
-            resized_file_path = f"{self.work_directory}/{prefix_splitted}-resized-{i}.jpg"
-            cv2.imwrite(resized_file_path, cv2.cvtColor(im, cv2.COLOR_RGB2BGR))
-            test_ds = SubsDataset(pd.DataFrame([{'path': resized_file_path}])['path'], pd.DataFrame([{'bb': np.array([0, 0, 0, 0])}])['bb'], pd.DataFrame([{'y': [0]}])['y'])
-            x, y_class, y_bb = test_ds[0]
-            xx = to_best_device(torch.FloatTensor(x[None, ]))
-            print(f"------ Inference")
-            out_class, out_bb = self.model(xx)
-            class_hat = torch.sigmoid(out_class.detach().cpu()).numpy()
-            if class_hat[0][0] >= self.threshold:
-                bb_hat = out_bb.detach().cpu()
-                bounding_boxes = get_bb_from_bouding_boxes(bb_hat, height=TARGET_HEIGHT, width=TARGET_WIDTH)
-                bb = bounding_boxes[0]
-                data.extend([[resized_file_path, True, bb[0], bb[1], bb[2], bb[3]]])
+        while True:
+            splitted_file = f"{self.work_directory}/{prefix_splitted}-{i}.jpg"
+            if os.path.exists(splitted_file):
+                print(f"--- Reading splitted file {splitted_file}")
+                im = read_image(splitted_file)
+                print(f"------ Resizing")
+                im = cv2.resize(im, size)
+                resized_file_path = f"{self.work_directory}/{prefix_splitted}-resized-{i}.jpg"
+                cv2.imwrite(resized_file_path, cv2.cvtColor(im, cv2.COLOR_RGB2BGR))
+                test_ds = SubsDataset(pd.DataFrame([{'path': resized_file_path}])['path'], pd.DataFrame([{'bb': np.array([0, 0, 0, 0])}])['bb'], pd.DataFrame([{'y': [0]}])['y'])
+                x, y_class, y_bb = test_ds[0]
+                xx = to_best_device(torch.FloatTensor(x[None, ]))
+                print(f"------ Inference")
+                out_class, out_bb = self.model(xx)
+                class_hat = torch.sigmoid(out_class.detach().cpu()).numpy()
+                if class_hat[0][0] >= self.threshold:
+                    bb_hat = out_bb.detach().cpu()
+                    bounding_boxes = get_bb_from_bouding_boxes(bb_hat, height=TARGET_HEIGHT, width=TARGET_WIDTH)
+                    bb = bounding_boxes[0].numpy()
+                    y0 = int(np.floor(min(bb[0], bb[2])))
+                    y1 = int(np.ceil(max(bb[0], bb[2])))
+                    x0 = int(np.floor(min(bb[1], bb[3])))
+                    x1 = int(np.ceil(max(bb[1], bb[3])))
+                    data.extend([[resized_file_path, True, x0, y0, x1, y1, class_hat[0][0]]])
+                else:
+                    data.extend([[resized_file_path, False, 0, 0, 0, 0, class_hat[0][0]]])
+                print(f"--- Splitted file {splitted_file} treated")
+                os.remove(splitted_file)
             else:
-                data.extend([[resized_file_path, False, 0, 0, 0, 0]])
-            print(f"--- Splitted file {splitted_file} treated")
-            os.remove(splitted_file)
+                print(f"--- Splitted file {splitted_file} does not exist so we will stop")
+                break
             i += 1
-        annotations = pd.DataFrame(columns=['filename', 'subs', 'x0', 'y0', 'x1', 'y1'], data=data)
+        annotations = pd.DataFrame(columns=['filename', 'subs', 'x0', 'y0', 'x1', 'y1', 'p'], data=data)
         annotations_file_path = f"{self.work_directory}/annotations_{prefix_splitted}.csv"
         annotations.to_csv(annotations_file_path, encoding='utf-8')
-        return annotations_file_path
+        return annotations_file_path, annotations
